removed middlemen for API functions
