+++ pages/api/refresh_token.ts
@@ -1,12 +1,24 @@
 import httpProxy from "http-proxy";
 import type { NextApiRequest, NextApiResponse } from "next";
 import { apiConfig } from "../../lib/apiFunctions/apiConfig";
-import refreshToken from "../../lib/apiFunctions/refreshToken";
+import {
+	forwardRequest,
+	getCommonRequirements,
+	handleRefresh,
+	handleResponse,
+	prepareForForwarding
+} from "../../lib/apiFunctions/commonFunctions";
+import { initCookies } from "../../lib/apiFunctions/responseHelpers";
 
 const proxy = httpProxy.createProxyServer({ changeOrigin: true });
 
 export const config = apiConfig;
 
 export default function route(req: NextApiRequest, res: NextApiResponse): Promise<void> {
-	return refreshToken(req, res, proxy);
+	return new Promise((resolve, reject): void => {
+		const { refreshToken } = getCommonRequirements(initCookies(req, res));
+		prepareForForwarding({ req, cookie: `jid=${refreshToken}` });
+		forwardRequest({ req, res, proxy, handleRes: true, reject });
+		handleResponse({ proxy, resolve, reject, handler: handleRefresh });
+	});
 }
